refactor data handler.
